Removed IOInput and IOOutput classes
Functionality moved to Storage class.
